feat: restructure package to export functionality
