+++ Makefile
@@ -1,0 +1,11 @@
+.PHONY: test
+test: mod-tidy
+	go test -v ./...
+
+.PHONY: mod-tidy
+mod-tidy:
+	go mod tidy
+
+.PHONY: build
+build: test
+	go build ./cmd/ccv
